refactor!:NativeUserLocation is deprecated use LocationPuck
