+++ example/src/examples/UserLocation/CustomNativeUserLocation.tsx
@@ -4,7 +4,7 @@ import {
   MapView,
   Camera,
   UserTrackingMode,
-  NativeUserLocation,
+  LocationPuck,
   Images,
   Image,
 } from '@rnmapbox/maps';
@@ -40,7 +40,7 @@ const UserLocationNativeAnimated = () => {
           followUserMode={UserTrackingMode.Follow}
           followZoomLevel={14}
         />
-        <NativeUserLocation
+        <LocationPuck
           topImage="topImage"
           visible={true}
           scale={['interpolate', ['linear'], ['zoom'], 10, 1.0, 20, 4.0]}
@@ -59,13 +59,9 @@ export default UserLocationNativeAnimated;
 
 const metadata: ExampleWithMetadata['metadata'] = {
   title: 'Custom Native UserLocation',
-  tags: [
-    'NativeUserLocation',
-    'NativeUserLocation#topImage',
-    'NativeUserLocation#scale',
-  ],
+  tags: ['LocationPuck', 'LocationPuck#topImage', 'LocationPuck#scale'],
   docs: `
-  Demonstrates use of images to customize NativeUserLocation
+  Demonstrates use of images to customize LocationPuck
   `,
 };
 UserLocationNativeAnimated.metadata = metadata;
